♻️refactor: 5주차 BaseEntity 경로 수정
