+++ src/main/java/com/project/likelion14thbe/domain/review/entity/Review.java
@@ -1,6 +1,6 @@
 package com.project.likelion14thbe.domain.review.entity;
 
-import com.project.likelion14thbe.domain.member.entity.BaseEntity;
+import com.project.likelion14thbe.global.BaseEntity;
 import com.project.likelion14thbe.domain.member.entity.Member;
 import com.project.likelion14thbe.domain.product.entity.Product;
 import jakarta.persistence.*;
